Proper handling of List<WebElement> + Visibility Group support
Lists evidently take a while to populate, which was resulting in NPEs
when calling .get(0) in the previous list handling code. Changed to use
the Selenium wait for lists.

Still no support for List<Link>, for example, though - need to
replicate TypifiedElement code for casting to a List<WebElement>.

Also added Visible(“Group Name”) support - long term idea being
this.waitForVisibilityOfGroup(“group that loads after an event”)
support for groups of elements that we want to wait for.